[Mod] #85 : Limiting effect of Copy/Paste part of code
New function is_initial_author take 2 params, the branch to be retrieved
and the kind of branch (hotfix / release).

With those params, the function check if the initial author of the branch
is same as current one, and display a Warning to alert the user that it
should be more precotious.

The option -I as been added to bypass the Yes/No question to continu retrieving
the branch